backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.4
python-dotenv==1.0.0
pillow==10.1.0
pyjwt==2.8.0
bcrypt==4.1.2
httpx==0.24.1
gotrue==1.3.0
